Roll back to v2.3.1
* Roll back to v2.3.1

+++ src/log.d
@@ -10,7 +10,7 @@ version(Notifications) {
 }
 
 // enable verbose logging
-long verbose;
+int verbose;
 bool writeLogFile = false;
 
 private bool doNotifications;
